requirements.txt
Django==3.2.9
django-celery-beat==2.3.0
django-celery-results==2.4.0
djangorestframework==3.13.1
psycopg2-binary==2.9.3
drf-yasg==1.20.0
django-silk==5.0.1
pdfkit==1.0.0
factory-boy==3.2.1